# Conflicts:
#	acf-json/group_5afd260eeb4ab.json
#	library/classes/process/class-woody-process.php

+++ library/classes/content/class-images.php
@@ -397,17 +397,25 @@ class WoodyTheme_Images
             $deleted_attachement = [];
         }
 
-        if (wp_attachment_is_image($attachment_id) && is_array($deleted_attachement) && !in_array($attachment_id, $deleted_attachement)) {
-            // Remove translations
-            $translations = pll_get_post_translations($attachment_id);
-            $deleted_attachement = array_merge($deleted_attachement, array_values($translations));
-            wp_cache_set('woody_deleted_attachement', $deleted_attachement, 'woody');
+        // Lors de la suppression d'une langue on doit supprimer tous ses médias pour éviter qu'ils ne passent dans la langue par défaut
+        // Pour cela on passe par une commande CLI et on ne veut surtout pas supprimer les traductions des médias supprimés
+        // On ne supprime pas les traductions d'une image si la suppression se fait en CLI
+        if (WP_CLI !== true) {
+            output_log('Attachment ' . $attachment_id . ' has been delted manualy');
+            if (wp_attachment_is_image($attachment_id) && is_array($deleted_attachement) && !in_array($attachment_id, $deleted_attachement)) {
+                // Remove translations
+                $translations = pll_get_post_translations($attachment_id);
+                $deleted_attachement = array_merge($deleted_attachement, array_values($translations));
+                wp_cache_set('woody_deleted_attachement', $deleted_attachement, 'woody');
 
-            foreach ($translations as $t_attachment_id) {
-                if ($t_attachment_id != $attachment_id) {
-                    wp_delete_attachment($t_attachment_id);
+                foreach ($translations as $t_attachment_id) {
+                    if ($t_attachment_id != $attachment_id) {
+                        wp_delete_attachment($t_attachment_id);
+                    }
                 }
             }
+        } else {
+            output_log('Attachment ' . $attachment_id . ' has been deleted by CLI');
         }
     }
